Magesh_stanch_project/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
python-dateutil==2.8.2
